itn-anchoring-handler/requirements.txt
pika
python-dotenv
azure-data-tables
pymssql